requirements.txt
customtkinter
Pillow
google-generativeai
python-dotenv
pandas
openpyxl
pdf2image
